pyproject.toml
[project]
name = "crypto-news-analyzer"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "beautifulsoup4>=4.14.2",
    "feedparser>=6.0.12",
    "python-dotenv>=1.2.1",
    "requests>=2.32.5",
    "schedule>=1.2.2",
    "sqlalchemy>=2.0.44",
    "trafilatura>=2.0.0",
]
